requirements.txt
absl-py==1.2.0
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.0
anyio==4.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.4.0
audioread==3.0.0
auto-gptq==0.2.2
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bitsandbytes-windows==0.37.5
bleach==5.0.0
blinker==1.6.2
cachetools==5.2.0
certifi==2023.7.22
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.7
clldutils==3.12.0
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.1.1
cryptography==41.0.4
csvw==3.1.1
cuda-python==11.7.1
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.14.5
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
diskcache==5.6.3
dnspython==2.4.2
docx2txt==0.8
emoji==2.8.0
entrypoints==0.4
epitran==1.23
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==0.8.3
faiss-cpu==1.7.4
Faker==19.6.2
fastapi==0.99.1
fastjsonschema==2.15.3
favicon==0.7.0
ffmpy==0.3.1
filelock==3.12.4
filetype==1.2.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.5.4
gitdb==4.0.9
GitPython==3.1.27
gradio==3.44.4
gradio_client==0.5.1
greenlet==2.0.2
h11==0.14.0
htbuilder==0.6.2
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.2
humanfriendly==10.0
idna==3.4
importlib-metadata==4.11.4
importlib-resources==6.1.0
InstructorEmbedding==1.0.1
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.6.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.5
langchain==0.0.267
langsmith==0.0.40
librosa==0.9.2
llama_cpp_python==0.2.6
llvmlite==0.39.1
lxml==4.9.3
Markdown==3.4.1
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.2
nltk==3.8.1
notebook==6.4.11
numba==0.56.2
numexpr==2.8.6
numpy==1.23.5
oauthlib==3.2.1
onnxruntime==1.16.0
openai-whisper==20230918
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.7
overrides==7.4.0
packaging==23.1
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
phonemizer==3.2.1
pickleshare==0.7.5
Pillow==9.1.1
pooch==1.6.0
posthog==3.0.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.1
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.0.0
Pygments==2.12.0
pymdown-extensions==10.3
pymongo==4.5.0
Pympler==1.0.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pypandoc==1.11
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
pyrsistent==0.18.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ffmpeg==2.0.4
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0.1
pyzmq==23.1.0
rdflib==6.2.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.4.4
rouge==1.0.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.1.2
scipy==1.11.2
segments==2.2.1
semantic-version==2.10.0
semver==2.13.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-rust==1.7.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.21
st-annotated-text==4.0.1
stack-data==0.2.0
starlette==0.27.0
streamlit==1.27.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.1
tensorflow-text==2.9.0
termcolor==2.3.0
terminado==0.15.0
threadpoolctl==3.2.0
tiktoken==0.3.3
tinycss2==1.1.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==2.0.1
torchvision==0.15.2
tornado==6.1
tqdm==4.66.1
traitlets==5.2.2.post1
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2022.1
tzlocal==4.2
unstructured==0.10.16
urllib3==1.26.6
uvicorn==0.23.2
validators==0.19.0
watchdog==2.1.8
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==3.6.0
wrapt==1.15.0
xxhash==3.3.0
yake==0.4.8
yarl==1.9.2
zipp==3.8.0
